+++ src/wrapping.rs
@@ -15,7 +15,7 @@
 
 use crate::{
     from_str::ParseFixedError,
-    traits::{Fixed, FixedSigned, FixedStrict, FixedUnsigned, FromFixed, ToFixed},
+    traits::{Fixed, FixedBits, FixedSigned, FixedStrict, FixedUnsigned, FromFixed, ToFixed},
     types::extra::{If, True},
     FixedI128, FixedI16, FixedI32, FixedI64, FixedI8, FixedU128, FixedU16, FixedU32, FixedU64,
     FixedU8,
@@ -1110,7 +1110,7 @@ impl<F: Fixed> Wrapping<F> {
     #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn mul_add<const MUL_FRAC: i32>(
         self,
-        mul: Wrapping<F::FixedF<MUL_FRAC>>,
+        mul: Wrapping<<F::Bits as FixedBits>::Fixed<MUL_FRAC>>,
         add: Wrapping<F>,
     ) -> Wrapping<F> {
         Wrapping(self.0.wrapping_mul_add(mul.0, add.0))
@@ -1140,8 +1140,8 @@ impl<F: Fixed> Wrapping<F> {
     #[inline]
     pub fn add_prod<const A_FRAC: i32, const B_FRAC: i32>(
         self,
-        a: Wrapping<F::FixedF<A_FRAC>>,
-        b: Wrapping<F::FixedF<B_FRAC>>,
+        a: Wrapping<<F::Bits as FixedBits>::Fixed<A_FRAC>>,
+        b: Wrapping<<F::Bits as FixedBits>::Fixed<B_FRAC>>,
     ) -> Wrapping<F> {
         Wrapping(self.0.wrapping_add_prod(a.0, b.0))
     }
@@ -1171,8 +1171,8 @@ impl<F: Fixed> Wrapping<F> {
     #[inline]
     pub fn mul_acc<const A_FRAC: i32, const B_FRAC: i32>(
         &mut self,
-        a: Wrapping<F::FixedF<A_FRAC>>,
-        b: Wrapping<F::FixedF<B_FRAC>>,
+        a: Wrapping<<F::Bits as FixedBits>::Fixed<A_FRAC>>,
+        b: Wrapping<<F::Bits as FixedBits>::Fixed<B_FRAC>>,
     ) {
         self.0.wrapping_mul_acc(a.0, b.0);
     }